[quant][refactor tests] Move qtensor serialization tests from test_deprecated_jit Summary: Move these tests into test_quantized_tensor
Test Plan:
python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
