+++ test/test_quantization.py
@@ -24,6 +24,7 @@ from quantization.core.test_quantized_functional import TestQuantizedFunctionalO
 from quantization.core.test_workflow_ops import TestFakeQuantizeOps  # noqa: F401
 # 3. Quantized Tensor
 from quantization.core.test_quantized_tensor import TestQuantizedTensor  # noqa: F401
+from quantization.core.test_quantized_tensor import TestQTensorSerialization  # noqa: F401
 # 4. Modules
 from quantization.core.test_workflow_module import TestFakeQuantize  # noqa: F401
 from quantization.core.test_workflow_module import TestObserver  # noqa: F401
